+++ pom.xml
@@ -22,7 +22,7 @@
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
     <version>2.2.0.RELEASE</version>
-    <relativePath/><!-- lookup parent from repository -->
+    <relativePath/><!-- lookup parent from repositories -->
   </parent>
 
   <dependencies>
@@ -33,6 +33,7 @@
     <dependency>
       <groupId>mysql</groupId>
       <artifactId>mysql-connector-java</artifactId>
+      <scope>runtime</scope>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
