+++ src/main/java/ru/geekbrains/gb/Application.java
@@ -4,9 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Applications {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(Applications.class);
+        SpringApplication.run(Application.class, args);
     }
 }
