--- src/main/java/ru/geekbrains/gb/ProductMapper.java
@@ -1,4 +1,0 @@
-package ru.geekbrains.gb;
-
-public class ProductMapper {
-}
