+++ src/main/java/ru/geekbrains/gb/models/ProductModel.java
@@ -1,0 +1,39 @@
+package ru.geekbrains.gb.models;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "products_items")
+public class ProductModel {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+
+    private Long id;
+    private String title;
+    private Integer cost;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public Integer getCost() {
+        return cost;
+    }
+
+    public void setCost(Integer cost) {
+        this.cost = cost;
+    }
+}
