+++ src/main/java/ru/geekbrains/gb/repositories/ProductRepository.java
@@ -1,0 +1,9 @@
+package ru.geekbrains.gb.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import ru.geekbrains.gb.models.ProductModel;
+
+
+public interface ProductRepository extends CrudRepository<ProductModel, Long> {
+
+}
