+++ src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -9,5 +9,10 @@
       "name": "spring.jpa.hibernate.naming-strategy",
       "type": "java.lang.String",
       "description": "Description for spring.jpa.hibernate.naming-strategy."
+    },
+    {
+      "name": "spring.java.properties.hibernate.dialect",
+      "type": "java.lang.String",
+      "description": "Description for spring.java.properties.hibernate.dialect."
     }
   ] }
