+++ src/main/resources/application.properties
@@ -1,13 +1,17 @@
+server.servlet.context-path=/products
 server.port=8080
-spring.datasource.driverClassName=com.mysql.jdbc.Driver
-spring.datasource.url=jdbc:mysql://${ MYSQL_HOST:localhost}:3306/springbootbase
+
+spring.mvc.formcontent.filter.enabled=true
+
+spring.datasource.url=jdbc:mysql://localhost:3306/springbootdatabase?createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true&useSSL=false&useUnicode=true&characterEncoding=UTF8&serverTimezone=UTC
 spring.datasource.username=root
 spring.datasource.password=1234
-spring.datasource.testOnBorrow=true
-spring.datasource.testWhileIdle=true
-spring.datasource.timeBetweenEvictionRunsMillis=60000
-spring.datasource.minEvictableIdleTimeMillis=30000
-spring.datasource.validationQuery=SELECT 1
-spring.datasource.max-active=15
-spring.datasource.max-idle=10
-spring.datasource.max-wait=8000
+spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
+spring.java.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.hibernate.ddl-auto=update
+
+logging.level.root=INFO
