--- src/main/resources/static/css/search_style.css
@@ -1,18 +1,0 @@
-form {
-    margin: 30px;
-}
-
-label {
-    font-family: Arial;
-    font-weight: bold ;
-    margin-right: 10px;
-}
-
-.btn{
-    font-family: Arial;
-    font-weight: bold;
-    color: azure;
-    background-color: cornflowerblue;
-    border-color: transparent;
-    border-radius: 10px;
-}
