--- src/main/resources/templates/product-search.html
@@ -1,16 +1,0 @@
-
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <link href="/css/search_style.css" rel="stylesheet">
-    <title>[ SEARCH ]</title>
-</head>
-<body>
-<form action="/product-search" method="POST">
-    <label>ID:</label>
-    <input type="text" name="id">
-    <input class="btn" type="submit" value="search">
-</form>
-</body>
-</html>
