+++ src/main/resources/templates/products.html
@@ -3,27 +3,17 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>[ SEARCH RESULT ]</title>
-    <link href="/css/result_style.css" rel="stylesheet">
+    <title>[ MAIN LIST ]</title>
+    <link href="/css/productlist_style.css" rel="stylesheet">
 </head>
 <body>
-<table>
-    <tr>
-        <th>ID</th>
-        <th>Title</th>
-        <th>Price</th>
-    </tr>
-    <th:block th:each="product : ${products}">
-        <tr>
-            <td th:text="${product.id}"></td>
-            <td th:text="${product.title}"></td>
-            <td th:text="${product.cost}"></td>
-        </tr>
-    </th:block>
-</table>
-<form action="/product-search" method="post" th:object="${product}">
-    <input name="ID" type="text">
-    <input class="btn" type="submit" value="search">
-</form>
+<h2>[ PRODUCTS LIST ]</h2>
+
+<div th:each="product : ${products} ">
+    <h4 th:text="${product.id}"/>
+    <h4 th:text="${product.title}"/>
+    <h4 th:text="${product.cost}"/>
+</div>
+
 </body>
 </html>
