--- src/main/webapp/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Main page</title>
-</head>
-<body>
-<p>Get your search <a href="/product-search">here</a></p>
-</body>
-</html>
